requirements.txt
playwright==1.48.0
aiogram==3.13.1
python-dotenv==1.0.1
requests==2.32.4
